client/{core,asset}: add asset.OrderCoin, refactor funding code
RedeemScript was previously removed from the asset.Coin interface as it is
only needed when funding orders, with FundOrder modified to return a slice
of redeem scripts with selected order funding coins.

This adds an asset.OrderCoin that includes the redeem script (where necessary)
to simplify the values returned by FundOrder and prevent potential bugs.

General refactor is also done to apply very similar and simplified logic and
code flow in the btc and dcr implementations of FundOrder, FundingCoins and
related methods.